Postmanをクラス化
